Refactored code and added support for reading from TSV files
Need to add some tests
