+++ spark-tensorflow/trainer/tests.py
@@ -14,24 +14,33 @@
 
 import argparse
 import tensorflow as tf
+
+import data
+import model
 import task
 
 ARTIFACT_DIR = './test/artifacts/'
 DATA_DIR = './test/'
-DATA_FILE = 'data.tfrecords'
 
 
-class TaskTests(tf.test.TestCase):
+class SampleTests(tf.test.TestCase):
   def test_get_feature_columns(self):
-    feature_columns = task.get_feature_columns(ARTIFACT_DIR)
+    feature_columns = data.get_feature_columns(
+      data.TSV,
+      ARTIFACT_DIR)
     self.assertEqual(len(feature_columns),
-                     len(task.INTEGER_FEATURES) + len(task.CATEGORICAL_FEATURES)
+                     len(data.INTEGER_FEATURES) + len(data.CATEGORICAL_FEATURES)
                      )
 
-  def test_generate_labelled_input_fn(self):
+  def test_generate_labelled_input_fn_tsv(self):
+    TSV_DATA_FILE = 'train.tsv'
     batch_size = 2
-    data_glob = '{}{}'.format(DATA_DIR, DATA_FILE)
-    labelled_input_fn = task.generate_labelled_input_fn(2, data_glob)
+    data_glob = '{}{}'.format(DATA_DIR, TSV_DATA_FILE)
+    labelled_input_fn = data.generate_labelled_input_fn(
+      data.TSV,
+      2,
+      data_glob,
+      ARTIFACT_DIR)
     features, labels = labelled_input_fn()
 
     with tf.Session() as sess:
@@ -49,11 +58,12 @@ class TaskTests(tf.test.TestCase):
     labels_out = result['labels']
     self.assertEqual(labels_out.shape, (batch_size,1))
 
-  def test_end_to_end(self):
+  def test_end_to_end_tsv(self):
     job_dir = tf.test.get_temp_dir()
 
     args = argparse.Namespace(
         job_dir=job_dir,
+        data_format=data.TSV,
         train_dir=DATA_DIR,
         eval_dir=DATA_DIR,
         artifact_dir=ARTIFACT_DIR,
@@ -64,26 +74,7 @@ class TaskTests(tf.test.TestCase):
         min_eval_frequency=0
     )
 
-    config = tf.contrib.learn.RunConfig()
-    params = tf.contrib.training.HParams(learning_rate=args.learning_rate)
-
-    estimator = task.generate_estimator(args.artifact_dir,
-                                        args.job_dir,
-                                        params,
-                                        config)
-
-    train_glob = '{}*.tfrecords'.format(args.train_dir)
-    eval_glob = '{}*.tfrecords'.format(args.eval_dir)
-
-    experiment = task.generate_experiment(estimator,
-                                          args.batch_size,
-                                          train_glob,
-                                          eval_glob,
-                                          args.train_steps,
-                                          args.eval_steps,
-                                          args.min_eval_frequency)
-
-    experiment.train_and_evaluate()
+    task.dispatch(args)
 
 if __name__ == '__main__':
   tf.test.main()
